Remove all unused functions during link
By default, gcc puts all functions within one source file in one section.
If a function is not used, the linker can later not throw out this function
until all other functions within the section (=the source file) are also not
used. With this patch each function will get it's own section, so the linker
can later throw out everything that is not used.

This is described e.g. here:
http://embeddedfreak.wordpress.com/2009/02/10/removing-unused-functionsdead-codes-with-gccgnu-ld/