ntp status: convert to clean multi-line output
Convert 'ntp status' to clean multi-line output with sequences of
ECMD_AGAIN() followed by a final ECMD_FINAL() to avoid having all
but the first line indented on e.g. serial connections that expect
"\r\n" for a line break.
In addition to that this fixes an overflow of the output buffer.

Signed-off-by: Peter Marschall <peter@adpm.de>
Signed-off-by: Christian Dietrich <stettberger@peer.zerties.org>